+++ py/desispec/fluxcalibration.py
@@ -592,9 +592,9 @@ def qa_fluxcalib(param, frame, fluxcalib):
     # Unpack model
 
     # Standard stars
-    stdfibers = (frame.fibermap['OBJTYPE'] == 'STD')
+    stdfibers = np.where((frame.fibermap['OBJTYPE'] == 'STD'))[0]
     stdstars = frame[stdfibers]
-    nstds = np.sum(stdfibers)
+    nstds = len(stdfibers)
     #try:
     #    assert np.array_equal(frame.fibers[stdfibers], input_model_fibers)
     #except AssertionError:
@@ -615,6 +615,7 @@ def qa_fluxcalib(param, frame, fluxcalib):
     # RMS
     qadict['NSTARS_FIBER'] = int(nstds)
     ZP_fiducial = np.zeros(nstds)
+
     for ii in range(nstds):
         # Good pixels
         gdp = stdstars.ivar[ii, :] > 0.
@@ -633,8 +634,8 @@ def qa_fluxcalib(param, frame, fluxcalib):
     imax = np.argmax(np.abs(ZPoffset))
     qadict['MAX_ZP_OFF'] = [float(ZPoffset[imax]),
                             int(stdfibers[np.argmax(ZPoffset)])]
-    if qadict['MAX_ZP_OFF'] > param['MAX_ZP_OFF']:
-        log.warn("Bad standard star ZP {:g}, in fiber {:d}".format(
+    if qadict['MAX_ZP_OFF'][0] > param['MAX_ZP_OFF']:
+        log.warning("Bad standard star ZP {:g}, in fiber {:d}".format(
                 qadict['MAX_ZP_OFF'][0], qadict['MAX_ZP_OFF'][1]))
     # Return
     return qadict